pledger: fix edge case in `-l`/`--last` date handling (#71)
The previous behavior used `Duration::weeks(4)` to calculate
the previous month, which wouldn't work on days in the month
beyond 28 (i.e., exactly 4 weeks).

The new approach fixes this by retrieving the previous month
symbolically and substituting it into the time. This should work
on all days of the month.